clean up code :tada:

+++ LodCore/LodCoreLibraryOld/Facades/UserManager.cs
@@ -3,31 +3,31 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Mail;
 using Journalist;
-using NHibernate.Util;
-using LodCoreLibraryOld.Domain.UserManagement;
-using LodCoreLibraryOld.Infrastructure.DataAccess.Repositories;
-using LodCoreLibraryOld.Domain.ProjectManagment;
-using LodCoreLibraryOld.Infrastructure.EventBus;
 using LodCoreLibraryOld.Common;
 using LodCoreLibraryOld.Domain.Exceptions;
+using LodCoreLibraryOld.Domain.ProjectManagment;
+using LodCoreLibraryOld.Domain.UserManagement;
+using LodCoreLibraryOld.Infrastructure.DataAccess.Repositories;
+using LodCoreLibraryOld.Infrastructure.EventBus;
+using NHibernate.Util;
 
 namespace LodCoreLibraryOld.Facades
 {
     public class UserManager : IUserManager
     {
         private readonly IConfirmationService _confirmationService;
-        private readonly IUserRepository _userRepository;
-        private readonly ProjectPaginationSettings _paginationSettings;
-        private readonly IProjectMembershipRepostiory _projectMembershipRepostiory;
-        private readonly IPasswordManager _passwordManager;
         private readonly IEventPublisher _eventPublisher;
+        private readonly ProjectPaginationSettings _paginationSettings;
+        private readonly IPasswordManager _passwordManager;
+        private readonly IProjectMembershipRepostiory _projectMembershipRepostiory;
+        private readonly IUserRepository _userRepository;
 
         public UserManager(
             IUserRepository userRepository,
-            IConfirmationService confirmationService, 
-            ProjectPaginationSettings paginationSettings, 
-            IProjectMembershipRepostiory projectMembershipRepostiory, 
-            ApplicationLocationSettings applicationLocationSettings, 
+            IConfirmationService confirmationService,
+            ProjectPaginationSettings paginationSettings,
+            IProjectMembershipRepostiory projectMembershipRepostiory,
+            ApplicationLocationSettings applicationLocationSettings,
             IPasswordManager passwordManager,
             IEventPublisher eventPublisher)
         {
@@ -54,13 +54,13 @@ namespace LodCoreLibraryOld.Facades
 
         public List<Account> GetUserList(int pageNumber, Func<Account, bool> criteria = null)
         {
-            var projectToSkip = _paginationSettings.NumberOfProjects*pageNumber;
+            var projectToSkip = _paginationSettings.NumberOfProjects * pageNumber;
             var projectsToTake = _paginationSettings.NumberOfProjects;
 
             return _userRepository.GetSomeAccounts(
-                projectToSkip, 
-                projectsToTake, 
-                account => account.RegistrationTime, 
+                projectToSkip,
+                projectsToTake,
+                account => account.RegistrationTime,
                 criteria);
         }
 
@@ -69,10 +69,7 @@ namespace LodCoreLibraryOld.Facades
             Require.Positive(userId, nameof(userId));
 
             var account = _userRepository.GetAccount(userId);
-            if (account == null)
-            {
-                throw new AccountNotFoundException();
-            }
+            if (account == null) throw new AccountNotFoundException();
 
             return account;
         }
@@ -91,10 +88,7 @@ namespace LodCoreLibraryOld.Facades
             Require.NotNull(request, nameof(request));
 
             var doesExist = EnumerableExtensions.Any(GetUserList(account => account.Email.Address == request.Email));
-            if (doesExist)
-            {
-                throw new AccountAlreadyExistsException();
-            }
+            if (doesExist) throw new AccountAlreadyExistsException();
 
             var newAccount = new Account(
                 request.Firstname,
@@ -131,10 +125,7 @@ namespace LodCoreLibraryOld.Facades
             Require.NotNull(account, nameof(account));
 
             var accountExists = _userRepository.GetAccount(account.UserId) != null;
-            if (!accountExists)
-            {
-                throw new AccountNotFoundException();
-            }
+            if (!accountExists) throw new AccountNotFoundException();
 
             _userRepository.UpdateAccount(account);
         }
@@ -147,19 +138,13 @@ namespace LodCoreLibraryOld.Facades
             var user = _userRepository.GetAccount(userId);
 
             var accountExists = user != null;
-            if (!accountExists)
-            {
-                throw new AccountNotFoundException();
-            }
+            if (!accountExists) throw new AccountNotFoundException();
 
             _passwordManager.UpdateUserPassword(userId, newPassword);
 
             var requestToDelete = _passwordManager.GetPasswordChangeRequest(userId);
 
-            if (requestToDelete != null)
-            {
-                _passwordManager.DeletePasswordChangeRequest(requestToDelete);
-            }
+            if (requestToDelete != null) _passwordManager.DeletePasswordChangeRequest(requestToDelete);
         }
 
         public void InitiatePasswordChangingProcedure(int userId)
@@ -189,7 +174,7 @@ namespace LodCoreLibraryOld.Facades
 
             var allUsersToSearchByRole =
                 new HashSet<Account>(
-                    allProjectMemberships.Select(membership => allUsers.Single(account => 
+                    allProjectMemberships.Select(membership => allUsers.Single(account =>
                         account.UserId == membership.DeveloperId)));
 
             var userRolesDictionary = allUsersToSearchByRole.ToDictionary(user => user,
@@ -198,8 +183,8 @@ namespace LodCoreLibraryOld.Facades
                         .Select(that => that.Role));
 
             return userRolesDictionary.Where(
-                pair =>
-                    pair.Value.Any(role => Extensions.Contains(role, searchString)))
+                    pair =>
+                        pair.Value.Any(role => Extensions.Contains(role, searchString)))
                 .Select(pair => pair.Key)
                 .Union(
                     allUsers.Where(
@@ -214,11 +199,8 @@ namespace LodCoreLibraryOld.Facades
             var account = _userRepository.GetAllAccounts().First(a => a.Email == new MailAddress(email));
 
             if (account.Password.Value == password)
-            {
                 return account;
-            }
-            else throw new Exception("Неверный пароль");
-
+            throw new Exception("Неверный пароль");
         }
     }
 }
